solved - cses/dpx3

+++ template.cpp
@@ -7,7 +7,7 @@ typedef vector<ll> vll;
 
 #define PB push_back
 
-const int MOD = 1e9 + 7;
+const long long MOD = 1e9 + 7;
 
 void solve() {
